% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFit.R
\name{checkFit}
\alias{checkFit}
\title{Check fit of nls estimate}
\usage{
checkFit(M, fit, N)
}
\arguments{
\item{M}{nls object (or NA)}

\item{fit}{the method used to determine the best fit; options: "n_abs" (default), "Pmean", "Psd", "Pn", "sigma"}

\item{N}{the actual (true) number of observations in the subset (numeric value)}
}
\value{
Returns a numeric value describing the fit, this should be minimized to choose a subset
}
\description{
Check fit of nls estimate
}
\details{
The parameter fit has the following options:
\itemize{
\item n_abs: the absolute deviation of the estimated number of observations in the subset to N
\item Pmean: the p-value of the parameter mean in the NLS model
\item Psd: the p-value of the parameter sd in the NLS model
\item Pn: the p-value of the parameter n in the NLS model
\item sigma: a parameter of the nls-function, which is the square root of the estimated variance of the random error
}
}
\seealso{
\code{\link{estimateCutoff}}
}
