% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIC.R
\docType{data}
\name{MIC}
\alias{MIC}
\title{Minimum Inhibitory Concentration Data}
\format{A data frame with 3340 rows and 24 variables:
\describe{
 \item{Antimicrobial}{factor, the name of the antibiotic}
 \item{Bacterium}{factor, the name of the bacterium}
\item{M0.002}{numeric, count of observations}
\item{M0.004}{numeric, count of observations}
\item{M0.008}{numeric, count of observations}
\item{M0.016}{numeric, count of observations}
\item{M0.032}{numeric, count of observations}
\item{M0.064}{numeric, count of observations}
\item{M0.125}{numeric, count of observations}
\item{M0.25}{numeric, count of observations}
\item{M0.5}{numeric, count of observations}
\item{M1}{numeric, count of observations}
\item{M2}{numeric, count of observations}
\item{M4}{numeric, count of observations}
\item{M8}{numeric, count of observations}
\item{M16}{numeric, count of observations}
\item{M32}{numeric, count of observations}
\item{M64}{numeric, count of observations}
\item{M128}{numeric, count of observations}
\item{M256}{numeric, count of observations}
\item{M512}{numeric, count of observations}
 \item{ECOFF}{numeric, the official ECOFF determined by EUCAST}
 \item{Distributions}{numeric, the number of sources for the observations}
 \item{Observations}{numeric, the total number of observations, so the sum of M0.002, ... ,M64}
}}
\source{
\url{https://mic.eucast.org/Eucast2/SearchController/search.jsp?action=init}
}
\usage{
MIC
}
\description{
The data give MIC distributions for individual organisms and
antimicrobial agents. The distributions are based on collated data from an increasing total of more than 20000 MIC
distributions from worldwide sources. Unless otherwise specifically  stated, the data are representative of results
obtained with a variety of MIC methods. Different methods do not give exactly the same results
but the results rarely vary by more than one doubling dilution step. In this way, the aggregated MIC distributions
encompass the variation between different investigators and between different methods.
}
\keyword{datasets}
