% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDistr.R
\name{plotDistr}
\alias{plotDistr}
\title{Visualization of the results of estimateCutoff}
\usage{
plotDistr(m, startP, DF)
}
\arguments{
\item{m}{matrix of all possible ECOFF estimates, the corresponding estimates for parameters mean, sd and n  and fit values}

\item{startP}{numeric vector of start parameters}

\item{DF}{data frame, containing columns "diam" and "obs" which specify the Zone Data}
}
\description{
Plotting the nls function with the estimated ECOFF.
Visualiziation through a pdf.
}
\seealso{
\code{\link{estimateCutoff}}
}
