% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeBacteria.R
\name{scrapeBacteria}
\alias{scrapeBacteria}
\title{Bacterial Data}
\source{
\url{https://mic.eucast.org/Eucast2/SearchController/search.jsp?action=init}
}
\usage{
scrapeBacteria(mic = T)
}
\arguments{
\item{mic}{logical; FALSE (default) Zone Data are generated}
}
\value{
Dowload a csv file with either MIC or Zone Data.
}
\description{
Generate and download a csv file for either Zone or Minimum Inhibitory Concentration (MIC) Data.
}
\details{
The function scrapeBacteria webscrapes the antimicrobial wild type distributions of microorganisms from the EUCAST-webpage.
It provides the collected data for all combinations of antibiotics and bacteria species in a CSV file for either MIC Data or Zone Data.

MIC Data:
\itemize{
 \item{Antimicrobial (the name of the antibiotic)}
 \item{Bacterium (the name of the bacterium)}
 \item{M0.002, ... ,M64 (19 variables, each of which contains the number of observations where the value specified in the name was observed)}
 \item{Distributions(the number of sources for the observations)}
 \item{Observations(the total number of observations, so the sum of M0.002, ... ,M64 )}
}

Zone Data:
\itemize{
 \item{Antimicrobial (the name of the antibiotic)}
 \item{Bacterium (the name of the bacterium)}
 \item{DiskContent (numeric)}
 \item{Z6, ... ,Z50 (45 variables, each of which contains the number of observations where the value (in millimeters) specified in the name was observed)}
 \item{ECOFF (the offcial Epidemiological Cutoff Value (short: ECOFF), which was determined by EUCAST)}
 \item{Distributions(the number of sources for the observations)}
 \item{Observations(the total number of observations, so the sum of M0.002, ... ,M64 )}
}
}
\seealso{
\code{\link{MIC}} \code{\link{ZD}}
}
