% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startParameters.R
\name{startParameters}
\alias{startParameters}
\title{Start Values for estimateCutoff}
\usage{
startParameters(DF, spMean = "mean")
}
\arguments{
\item{DF}{data frame, containing columns "diam" and "obs" which specify the Zone Data}

\item{spMean}{method for estimating mean start parameter;  "mean" (default) or "peak1"}
}
\value{
The function \code{startParameters} returns a numeric vector of estimated start parameters for mean, sd and n.
}
\description{
A heuristic function to estimate start values for the ECOFFinder.
}
\seealso{
\code{\link{estimateCutoff}}
}
