% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZD.R
\docType{data}
\name{ZD}
\alias{ZD}
\title{Zone Data}
\format{A data frame with 1626 rows and 51 variables:
\describe{
 \item{Antimicrobial}{factor, the name of the antibiotic}
 \item{Bacterium}{factor, the name of the bacterium}
 \item{DiskContent}{numeric}
\item{Z6}{numeric, count of observations}
\item{Z7}{numeric, count of observations}
\item{Z8}{numeric, count of observations}
\item{Z9}{numeric, count of observations}
\item{Z10}{numeric, count of observations}
\item{Z11}{numeric, count of observations}
\item{Z12}{numeric, count of observations}
\item{Z13}{numeric, count of observations}
\item{Z14}{numeric, count of observations}
\item{Z15}{numeric, count of observations}
\item{Z16}{numeric, count of observations}
\item{Z17}{numeric, count of observations}
\item{Z18}{numeric, count of observations}
\item{Z19}{numeric, count of observations}
\item{Z20}{numeric, count of observations}
\item{Z21}{numeric, count of observations}
\item{Z22}{numeric, count of observations}
\item{Z23}{numeric, count of observations}
\item{Z24}{numeric, count of observations}
\item{Z25}{numeric, count of observations}
\item{Z26}{numeric, count of observations}
\item{Z27}{numeric, count of observations}
\item{Z28}{numeric, count of observations}
\item{Z29}{numeric, count of observations}
\item{Z30}{numeric, count of observations}
\item{Z31}{numeric, count of observations}
\item{Z32}{numeric, count of observations}
\item{Z33}{numeric, count of observations}
\item{Z34}{numeric, count of observations}
\item{Z35}{numeric, count of observations}
\item{Z36}{numeric, count of observations}
\item{Z37}{numeric, count of observations}
\item{Z38}{numeric, count of observations}
\item{Z39}{numeric, count of observations}
\item{Z40}{numeric, count of observations}
\item{Z41}{numeric, count of observations}
\item{Z42}{numeric, count of observations}
\item{Z43}{numeric, count of observations}
\item{Z44}{numeric, count of observations}
\item{Z45}{numeric, count of observations}
\item{Z46}{numeric, count of observations}
\item{Z47}{numeric, count of observations}
\item{Z48}{numeric, count of observations}
\item{Z49}{numeric, count of observations}
\item{Z50}{numeric, count of observations}
 \item{ECOFF}{numeric, the official ECOFF determined by EUCAST}
 \item{Distributions}{numeric, the number of sources for the observations}
 \item{Observations}{numeric, the total number of observations, so the sum of M0.002, ... ,M64}
}}
\source{
\url{https://mic.eucast.org/Eucast2/SearchController/search.jsp?action=init}
}
\usage{
ZD
}
\description{
The data give inhibition zone diameter distributions for individual organisms and antimicrobial agents in tables and histograms. The
distributions are based on collated data from an increasing number of sources worldwide. The data are representative of results obtained
with the EUCAST disk diffusion method (launched in 2009).
}
\keyword{datasets}
